Modified the makefile to support experimental optional C++ compilation:  - new option CPP_COMPILE to trigger this (off by default)  - split CFLAGS into common, C-only, and C++-only flags  - when enabled, CPP_COMPILE causes 'pp' to be appended to     the target name
NOTE THAT THE SYSTEM CANNOT ACTUALLY BE COMPILED THIS WAY 
YET. IT IS JUST AN EXPERIMENT.

Modified lib.mak to always build zlib/expat as C regardless 
of CPP_COMPILE.

Modified windows.mak to fix warnings with MAXOPT=1, and to 
leverage the new CFLAGs definitions.

Modified vconv.c to do appropriate conversions for new C++ 
options.

Updated sources so that libutil, libocore (Windows), and 
libosd (Windows) can be cleanly compiled as C or C++. This
was mostly adding some casts against void *.

Fixed a few more general obvious problems at random 
locations in the source:
 - device->class is now device->devclass
 - TYPES_COMPATIBLE uses typeid() when compiled for C++
 - some functions with reserved names ('xor' in particular)
    were renamed
 - nested enums and structs were pulled out into separate
    definitions (under C++ these would need to be scoped to
    be referenced)
 - TOKEN_VALUE cannot use .field=x initialization in C++ :(

+++ src/emu/cpu/tms34010/34010ops.h
@@ -43,7 +43,7 @@ INLINE void TMS34010_WRMEM_DWORD(tms34010_state *tms, offs_t A,UINT32 V)
 
 /* IO registers accessor */
 #define IOREG(T,reg)				((T)->IOregs[reg])
-#define SMART_IOREG(T,reg)			((T)->IOregs[(T)->is_34020 ? REG020_##reg : REG_##reg])
+#define SMART_IOREG(T,reg)			((T)->IOregs[(T)->is_34020 ? (int)REG020_##reg : (int)REG_##reg])
 #define PBH(T) 						(IOREG(T, REG_CONTROL) & 0x0100)
 #define PBV(T) 						(IOREG(T, REG_CONTROL) & 0x0200)
